interpreter is now back to being compiled in C, with interface for running debugger, which is C++. Debugger code moved to interp/dbg. Variable printing updated accordingly.
