+++ interp/dbg/abstract_syntax_tree.cpp
@@ -9,26 +9,25 @@ abstract_syntax_tree::abstract_syntax_tree( const string& filename )
 	load( doc );
 }
 
-variable* abstract_syntax_tree::var( const string & func_name, const string & var_name )
+variable* abstract_syntax_tree::var( const string& func_name, const string & var_name )
 {
-	auto func_iter = _functions.find( func_name );
-	if( func_iter == _functions.end() )
-		return nullptr; //function doesnt exist
-	else
+    auto func_iter = _functions.find( func_name );
+    if( func_iter == _functions.end() )
+        return nullptr; //function doesnt exist
+    else
+    {
+	variable* var = func_iter->second.var( var_name );
+	if( var != nullptr )
+	    return var; //variable is in function
+	else //variable isn't in function
 	{
-		variable* var = func_iter->second.var( var_name );
-		if( var != nullptr )
-			return var; //variable is in function
-		else //variable isn't in function
-		{
-			auto var_iter = _globals.find( var_name );
-			if( var_iter != _globals.end() )
-				return &var_iter->second; //variable isnt in globals
-			return nullptr; //variable not in locals or globals
-		}
-	}
-	//return nullptr; //it's literally impossible to get here. Look at the 'else' path.
-	//dumb visual studio warnings.
+	    auto var_iter = _globals.find( var_name );
+	    if( var_iter != _globals.end() )
+		return &var_iter->second; //variable isnt in globals
+            else return nullptr; //variable not in locals or globals
+        }
+    }
+    return nullptr; 
 }
 
 void abstract_syntax_tree::load( xml_document<char>& doc )
