+++ interp/dbg/asm_list.h
@@ -16,11 +16,12 @@ using std::unordered_map;
 class asm_list
 {
 public:
+        asm_list() {} 
 	asm_list( const string& filename );
 
 	uint32_t addr_of_func( const string& func );
 	uint32_t addr_of_line( const string& filename, uint32_t line_number );
-	uint32_t line_of_addr( const string& filename, uint32_t line_number );
+	uint32_t line_of_addr( const string& filename, uint32_t address  );
 
 	string current_func( uint32_t cur_addr );
 	string current_file( uint32_t cur_addr );
