+++ interp/dbg/string_utils.h
@@ -14,4 +14,6 @@ public:
 	static vector<string> string_split( const string& s, const char delimiter );
 	static void ltrim( string &s );
 	static bool is_number( const string& num_text, int base = 10, int* res = nullptr );
+	static bool begins_with( const string& input, const string& begin );
+	static bool ends_with( const string& input, const string& end );
 };
