+++ interp/dbg/struct_decl.cpp
@@ -17,10 +17,11 @@ struct_decl::struct_decl( xml_node<char>* struct_node, unordered_map<string, str
 	}
 }
 
-string struct_decl::to_string( uint32_t fp ) const
+string struct_decl::to_string( Machine_State* cpu ) const
 {
 	string ret = "";
 	for( auto& var : _fields )
-		ret += "\t" + var.to_string( fp ) + '\n';
+		ret += "\t" + var.to_string( cpu ) + '\n';
+        ret += "}";
 	return ret;
 }
