+++ interp/dbg/struct_decl.h
@@ -12,6 +12,7 @@ using std::string;
 using rapidxml::xml_node;
 
 #include "variable.h"
+#include "../machine.h"
 
 class struct_decl
 {
@@ -19,12 +20,12 @@ public:
 	struct_decl( xml_node<char>* decl_node, unordered_map<string, struct_decl>& live_type_context );
 	struct_decl() {}
 
-	string to_string( uint32_t fp ) const;
+	string to_string( Machine_State* cpu ) const;
 	inline int size() const { return _size; }
 	inline string name() const { return _name; }
 
 private:
-	vector<variable> _fields; //key = field name, value = variable obj
+	vector<variable> _fields;
 	int _size = 0;
 	string _name;
 };
